api/src/vms/core/security.py
"""Utilitários de segurança: JWT, API keys, HMAC."""

import hashlib
import hmac
import secrets
import uuid
from datetime import UTC, datetime, timedelta

import bcrypt
from jose import JWTError, jwt

from vms.core.config import get_settings

# Algoritmo JWT
_ALGORITHM = "HS256"

# Prefixo de API keys — facilita identificação em logs
_API_KEY_PREFIX = "vms_"


# ─── Senhas ───────────────────────────────────────────────────────────────────

def hash_password(plain: str) -> str:
    """Gera hash bcrypt da senha."""
    return bcrypt.hashpw(plain.encode(), bcrypt.gensalt()).decode()


def verify_password(plain: str, hashed: str) -> bool:
    """Verifica senha contra hash bcrypt."""
    return bcrypt.checkpw(plain.encode(), hashed.encode())


# ─── JWT ─────────────────────────────────────────────────────────────────────

def create_access_token(
    subject: str,
    tenant_id: str,
    role: str,
    expire_minutes: int | None = None,
) -> str:
    """Emite access token JWT com claims de identidade."""
    settings = get_settings()
    expire_delta = expire_minutes or settings.access_token_expire_minutes

    payload = {
        "sub": subject,
        "tenant_id": tenant_id,
        "role": role,
        "type": "access",
        "exp": datetime.now(UTC) + timedelta(minutes=expire_delta),
        "iat": datetime.now(UTC),
        "jti": str(uuid.uuid4()),
    }
    return jwt.encode(payload, settings.secret_key, algorithm=_ALGORITHM)


def create_refresh_token(subject: str, tenant_id: str) -> str:
    """Emite refresh token JWT de longa duração."""
    settings = get_settings()

    payload = {
        "sub": subject,
        "tenant_id": tenant_id,
        "type": "refresh",
        "exp": datetime.now(UTC) + timedelta(days=settings.refresh_token_expire_days),
        "iat": datetime.now(UTC),
        "jti": str(uuid.uuid4()),
    }
    return jwt.encode(payload, settings.secret_key, algorithm=_ALGORITHM)


def decode_token(token: str) -> dict:
    """Decodifica e valida JWT. Lança JWTError se inválido."""
    settings = get_settings()
    return jwt.decode(token, settings.secret_key, algorithms=[_ALGORITHM])


def is_token_valid(token: str, expected_type: str = "access") -> bool:
    """Retorna True se o token for válido e do tipo esperado."""
    try:
        payload = decode_token(token)
        return payload.get("type") == expected_type
    except JWTError:
        return False


# ─── API Keys ─────────────────────────────────────────────────────────────────

def generate_api_key() -> tuple[str, str, str]:
    """
    Gera uma nova API key.

    Retorna:
        tuple: (plain_key, key_hash, prefix)
        - plain_key: valor completo, mostrado UMA vez ao usuário
        - key_hash: hash bcrypt para armazenar no banco
        - prefix: primeiros 12 chars (para lookup)
    """
    raw = secrets.token_urlsafe(32)
    plain_key = f"{_API_KEY_PREFIX}{raw}"
    prefix = plain_key[:12]
    key_hash = bcrypt.hashpw(plain_key.encode(), bcrypt.gensalt()).decode()
    return plain_key, key_hash, prefix


def verify_api_key(plain: str, hashed: str) -> bool:
    """Verifica API key contra hash armazenado."""
    return bcrypt.checkpw(plain.encode(), hashed.encode())


def extract_key_prefix(plain_key: str) -> str:
    """Extrai o prefixo de busca de uma API key."""
    return plain_key[:12]


# ─── HMAC para webhooks de saída ──────────────────────────────────────────────

def sign_webhook_payload(body: bytes, secret: str) -> str:
    """
    Gera assinatura HMAC-SHA256 para webhook de saída.

    O receptor deve verificar com:
        expected = sign_webhook_payload(body, shared_secret)
        hmac.compare_digest(received_signature, expected)
    """
    return hmac.new(
        secret.encode("utf-8"),
        body,
        hashlib.sha256,
    ).hexdigest()


def verify_webhook_signature(body: bytes, secret: str, signature: str) -> bool:
    """Verifica assinatura HMAC-SHA256 recebida de forma segura (timing-safe)."""
    expected = sign_webhook_payload(body, secret)
    return hmac.compare_digest(expected, signature)
